+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CourseRestController.java
@@ -1,8 +1,8 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.application;
 
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.CourseDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.controllers.CourseController;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/courses")
@@ -11,5 +11,10 @@ public class CourseRestController {
 
     public CourseRestController(CourseController courseController) {
         this.courseController = courseController;
+    }
+
+    @PostMapping("/{schoolId}")
+    public void createCourse(@RequestBody CourseDTO courseDTO, @PathVariable Long schoolId){
+        courseController.createCourse(courseDTO, schoolId);
     }
 }