+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/SchoolController.java
@@ -46,4 +46,8 @@ public class SchoolController {
         school.setCourseList(schoolDTO.getCourseList());
         return schoolService.saveSchool(school);
     }
+    public void deleteSchool(Long schoolId) {
+        schoolService.deleteSchool(schoolId);
+    }
+
 }