+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/persistence/StudentRepository.java
@@ -1,9 +1,0 @@
-package com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence;
-
-import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
-}