+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/CourseService.java
@@ -1,5 +1,8 @@
 package com.nimio2024.nimio2024_tfg_polmorenocastelar.service;
 
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.CourseDTO;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.School;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence.CourseRepository;
 import org.springframework.stereotype.Service;
 
@@ -9,5 +12,12 @@ public class CourseService {
 
     public CourseService(CourseRepository courseRepository) {
         this.courseRepository = courseRepository;
+    }
+
+    public Course createCourse(CourseDTO courseDTO, School school) {
+        Course course = new Course(courseDTO);
+        course.setSchool(school);
+        courseRepository.save(course);
+        return course;
     }
 }