+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/SchoolService.java
@@ -18,4 +18,6 @@ public class SchoolService {
     }
 
     public School getSchoolById(Long schoolId) {return schoolRepository.findById(schoolId).get();}
+    public void deleteSchool(Long schoolId) {schoolRepository.deleteById(schoolId);}
+
 }